refactor: rename V4 modules and extract o_proj for cleaner PR diff
- Delete stale docs/deepseek-v4-flash-kunlun.md (debug log, no longer useful)
- Rename mhc.py -> hyper_connection.py (clearer semantics)
- Rename deepseek_v4_topk.py -> deepseek_v4_sparse_index.py
- Extract deepseek_v4_bf16_o_proj from golden-owned deep_gemm.py into
  new deepseek_v4_o_proj.py to avoid merge conflicts on PR
- Update all import references in __init__.py
