+++ vllm_kunlun/ops/hyper_connection.py
